+++ kernel/src/arch/x86/cpuid.c
@@ -1,5 +1,5 @@
-#include "arch/x86/cpuid.h"
-#include "kstring.h"
+#include "kernel/include/arch/x86/cpuid.h"
+#include "util/kstdlib/include/kstring.h"
 
 /*
  * It's implemented in cpuid.s
@@ -9,9 +9,15 @@ uint32 CPUID_IsSuported();
 /*
  * It's implemented in cpuid.s
  */
-void CPUID_Call(uint32 eax_value, uint32 ebx_value,
-                uint32 ecx_value, uint32 edx_value,
-                uint32 *eax, uint32 *ebx, uint32 *ecx, uint32 *edx);
+void CPUID_Call(
+    uint32 eax_value, 
+    uint32 ebx_value,
+    uint32 ecx_value, 
+    uint32 edx_value,
+    uint32 *eax, 
+    uint32 *ebx, 
+    uint32 *ecx, 
+    uint32 *edx);
 
 struct CPUID_eax_01h
 {
@@ -42,7 +48,7 @@ struct CPUInfo
 
 static struct CPUInfo g_cpuInfo;
 
-size_t CPUID_Init()
+uint32 CPUID_Init()
 {
     size_t error = kmemset((void*) &g_cpuInfo, 0, sizeof(struct CPUInfo));
     if (error != ERROR_SUCCESS)
@@ -76,7 +82,8 @@ size_t CPUID_Init()
     return ERROR_SUCCESS;
 }
 
-size_t CPUID_GetVendorName(const char **a_vendorName)
+uint32 CPUID_GetVendorName(
+    const char **a_vendorName)
 {
     if (a_vendorName == NULL)
     {
@@ -93,7 +100,8 @@ size_t CPUID_GetVendorName(const char **a_vendorName)
     return ERROR_SUCCESS;
 }
 
-size_t CPUID_HasFeature(size_t a_featureId)
+uint32 CPUID_HasFeature(
+    size_t a_featureId)
 {
     if (g_cpuInfo.isInitialized == 0)
     {
